Turn off pluralize list titles
The hugo configuration settings were _under_ the params section
which in TOML means that they are part of params,
so hugo was completely ignoring them all.

I have moved them up a bit in the file so that they are parsed as
root level values in the file. Now the pluralizeListTitles setting
(along with a few others) will be applied by Hugo properly.