Formats all docs to use same wrapping
There was a wide variety of wrapping formats from 80 chars all the
way up to never wrapping. This changes the wrapping of all documents
to be at/around 80 chars to make it easier to review and read in text
editors.

I made no other changes to content save for sorting the list of Kubernetes
distros alphabetically to avoid anything that looks like preference for
any given distro.

Signed-off-by: Taylor Thomas <taylor.thomas@microsoft.com>