Clarify valid chart naming scheme
As noted in issue [#8012](https://github.com/helm/helm/issues/8012), the
ValidName regex have been incorrect for a while. While one could not
install charts with uppercase letters in the names, one did not get a
lint warning. I believe that is fixed since
[PR #8013](https://github.com/helm/helm/pull/8013).

This change clarifies that in the documentation.

Signed-off-by: Andreas Lindhé <andreas@lindhe.io>