Removing current banner that has been up for past 9 months
Note, the template has been updated so that the banner is optional
and the region only displays if a banner is present.

The banner was discussed among some org maintainers and then taken
to the org maintainers list. This change is after the normal lazy
concensus period and there was no disucssion.

Taking down the banner was raised as a question about when should
something like this happen? Is it when the problem is solved, when
another issue comes along to replace it, or some other time?

Given that other projects (e.g., Kubernetes) have removed their
banner and that Helm is not a political or activist organization
(so we don't see a next replacement banner along these lines)
now appears to be a reasonable time to remove it.

It is worth noting that the Helm community is a global community.

This is not to say that members of the Helm community do not care
about BLM.

Signed-off-by: Matt Farina <matt.farina@suse.com>