Provide better explanations of dot context
- See https://github.com/helm/helm/issues/11232 and
https://github.com/helm/helm/issues/7915 for examples of confusion the
current explanation/wording has caused.
- Replaced inconsistent and unexplained terminology: "root context",
"root scope", "top-level scope", "parent scope".
- Go templates/gomplate, on which Helm relies, uses the words "dot" and
"context", not "scope". The word "scope" is better used to explain when
variables go out of scope, not to explain the value of the context
variables `.` and `$`.
- Fixed inaccurate or misleading statements such as:
 $ "does not change", and $ "will always point to the root context".
- This sentence: "Recall that `.` is a reference to _the current
scope_." is not useful - this was not explained elsewhere
- Removed vague sentences that don't add meaning, such as "Values are
namespaced, but namespaces are pruned."

Signed-off-by: Shannon Carey <rehevkor5@gmail.com>