Fix web site scrolling on mobile browsers
Scrolling pages on the Helm web site is difficult on mobile devices due
to nested scrolling containers. This changeset removes the nested
scrolling containers.

The fix to remove the nested scrolling containers caused pages to start
scrolling horizontally because the content of the pages have negative
margins, which makes the page wider than 100%. The rest of this
changeset addresses the negative margins by removing them and/or
offsetting them in each element that was causing the page to horizontal
scroll.

Fixes #741
Fixes #1393
Fixes #1464

Signed-off-by: Matt Light <matt@tinyprint.dev>