Adding redirect for helm/v4 to the code
Helm uses custom URLs to get to the code. In this case helm.sh/helm/4
should point to the main branch for helm so that people who try to use
it with go modules are able to.

This change adds that redirect to the default language which means it
should work for everyone.

Signed-off-by: Matt Farina <matt@mattfarina.com>