Removing inactive maintainers from the owners
According to the governance maintainers must remain actice and
are removed for inactivity.

The maintainers moved to emeritus here have already signaled they
are going emeritus and this repo was missed or have been inactive
for more than 1 year.

Note, bridgetkromhout was moved in the ordering to keep the list
in alpha order.

Signed-off-by: Matt Farina <matt.farina@suse.com>