Update changelog script to node and update changelog with automated categories
Changelog script was missing items due to GitHub graphql items limit (limit 1000 but 1200+ records)
Now script is very slow, but this is the only reliable way to get all the PRs that include
cherry-picked commits, which GitHub API or UI do not detect as connected to the picked SHA.

Converting to node sped up the GitHub API calls from 14 mins to ~30 seconds with parallelism!

In order for auto-categorization to work, I manually updated labels for all PRs merged into the RC
that were not in v3.19.0, as well as their corresopnding v3 backport PRs.

Signed-off-by: Scott Rigby <scott@r6by.com>
Signed-off-by: Nicolas Fantoni <nfantoni@imolainformatica.it>