Make docs tags more accessible
We aren't currently using docs tags, but this could be a good way to categorize
docs content in addition to the sidebar.

This PR swizzles two components:
- the Docs/Footer, in order to remove the inline tags
- the Docs/Layout, in order to add the inline tags below the version badge

Orthogonal to https://github.com/helm/helm-www/issues/1864, but could enhance
that effort by providing links to pages that group certain types of docs
content together in various ways.

Signed-off-by: Scott Rigby <scott@r6by.com>